LPS-54704 Move helper tests classes from portal to dynamic-data-lists-test
